split up query_subset
