+++ engine/crates/partial-caching/src/query_subset/display.rs
@@ -1,113 +1,17 @@
 use std::fmt;
 
 use cynic_parser::{
-    executable::{
-        ids::{FragmentDefinitionId, OperationDefinitionId, SelectionId, VariableDefinitionId},
-        iter::{IdIter, Iter},
-        Selection, VariableDefinition,
-    },
+    executable::{ids::SelectionId, iter::Iter, Selection},
     ExecutableDocument,
 };
 use indexmap::IndexSet;
 
-/// Part of a query that was submitted to the API.
-///
-/// This is a group of fields with the same cache settings, and all the
-/// ancestors, variables & fragments required for those fields to make a
-/// valid query
-pub struct QuerySubset {
-    pub(crate) operation: OperationDefinitionId,
-    partition: Partition,
-    variables: IndexSet<VariableDefinitionId>,
-}
-
-#[derive(Default, Debug)]
-pub(crate) struct Partition {
-    pub selections: IndexSet<SelectionId>,
-    pub fragments: IndexSet<FragmentDefinitionId>,
-}
-
-impl QuerySubset {
-    pub(crate) fn new(
-        operation: OperationDefinitionId,
-        cache_group: Partition,
-        variables: IndexSet<VariableDefinitionId>,
-    ) -> Self {
-        QuerySubset {
-            operation,
-            partition: cache_group,
-            variables,
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.partition.selections.is_empty()
-    }
-
-    pub fn extend(&mut self, other: &QuerySubset) {
-        self.partition
-            .selections
-            .extend(other.partition.selections.iter().copied());
-        self.partition
-            .fragments
-            .extend(other.partition.fragments.iter().copied());
-        self.variables.extend(other.variables.iter().cloned());
-    }
-
-    pub fn as_display<'a>(&'a self, document: &'a ExecutableDocument) -> QuerySubsetDisplay<'a> {
-        QuerySubsetDisplay {
-            subset: self,
-            document,
-            include_query_name: false,
-        }
-    }
-
-    pub fn variables<'a>(
-        &'a self,
-        document: &'a ExecutableDocument,
-    ) -> impl Iterator<Item = VariableDefinition<'a>> + 'a {
-        self.variables.iter().map(|id| document.read(*id))
-    }
-
-    fn selection_set_display<'a>(&'a self, selections: Iter<'a, Selection<'a>>) -> SelectionSetDisplay<'a> {
-        SelectionSetDisplay {
-            visible_selections: &self.partition.selections,
-            selections: self.selection_iter(selections),
-            indent_level: 0,
-        }
-    }
-
-    pub(crate) fn selection_iter<'a>(&'a self, selection_set: Iter<'a, Selection<'a>>) -> FilteredSelections<'a> {
-        FilteredSelections {
-            visible_ids: &self.partition.selections,
-            selections: selection_set.with_ids(),
-        }
-    }
-}
-
-#[derive(Clone, Copy)]
-pub(crate) struct FilteredSelections<'a> {
-    visible_ids: &'a IndexSet<SelectionId>,
-    selections: IdIter<'a, Selection<'a>>,
-}
-
-impl<'a> Iterator for FilteredSelections<'a> {
-    type Item = Selection<'a>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        for (id, selection) in self.selections.by_ref() {
-            if self.visible_ids.contains(&id) {
-                return Some(selection);
-            }
-        }
-        None
-    }
-}
+use super::{FilteredSelectionSet, QuerySubset};
 
 pub struct QuerySubsetDisplay<'a> {
-    subset: &'a QuerySubset,
-    document: &'a ExecutableDocument,
-    include_query_name: bool,
+    pub(super) subset: &'a QuerySubset,
+    pub(super) document: &'a ExecutableDocument,
+    pub(super) include_query_name: bool,
 }
 
 impl QuerySubsetDisplay<'_> {
@@ -119,10 +23,10 @@ impl QuerySubsetDisplay<'_> {
     }
 }
 
-struct SelectionSetDisplay<'a> {
-    selections: FilteredSelections<'a>,
-    visible_selections: &'a IndexSet<SelectionId>,
-    indent_level: usize,
+pub(super) struct SelectionSetDisplay<'a> {
+    pub(super) selections: FilteredSelectionSet<'a, 'a>,
+    pub(super) visible_selections: &'a IndexSet<SelectionId>,
+    pub(super) indent_level: usize,
 }
 
 struct SelectionDisplay<'a> {
@@ -135,7 +39,7 @@ impl<'a> SelectionDisplay<'a> {
     fn wrap_set(&self, selections: Iter<'a, Selection<'a>>) -> SelectionSetDisplay<'a> {
         SelectionSetDisplay {
             visible_selections: self.visible_selections,
-            selections: FilteredSelections {
+            selections: FilteredSelectionSet {
                 visible_ids: self.visible_selections,
                 selections: selections.with_ids(),
             },
